Migrate sensors to extend SensorEntity.
Temp, humidity, power and energy sensors are now marked as measured.
This is to support long-term statistics.

The energy sensors are now marked as an energy class instead of power.
The 'total today energy' has a new property; last_reset. This is used by
the build-in Energy Management to keep track of the energy consumption.

The 'total today energy' will not pass a 'period' to the Ngenic API,
doing so lead to missing meterings during the day.